// frontend/src/api/schemas/appointmentDto.ts
/**
 * Generated by orval v8.9.0 🍺
 * Do not edit manually.
 * Clic Flow API
 * API documentation for Clic Flow
 * OpenAPI spec version: 1.0
 */

export interface AppointmentDto {
  id: string;
  clinicId: string;
  patientId: string;
  doctorId: string;
  date: string;
  appointmentPriceInCents: number;
  createdAt: string;
  updatedAt: string;
}