// frontend/src/api/schemas/patientDto.ts
/**
 * Generated by orval v8.9.0 🍺
 * Do not edit manually.
 * Clic Flow API
 * API documentation for Clic Flow
 * OpenAPI spec version: 1.0
 */
import type { PatientDtoSex } from "./patientDtoSex";

export interface PatientDto {
  id: string;
  clinicId: string;
  name: string;
  email: string;
  phoneNumber: string;
  sex: PatientDtoSex;
  createdAt: string;
  updatedAt: string;
}